feat: updates api sample and fills the result interface using reflection instead of json package

+++ api/model.go
@@ -1,5 +1,5 @@
 package common
 
 type Data struct {
-	Data float64
+	Data int
 }
